moved srv.py to lib/
